Refactor: new 'shop' folder for shop.php support
