+++ src/net/sourceforge/kolmafia/textui/command/CoinmasterCommand.java
@@ -8,9 +8,9 @@ import net.sourceforge.kolmafia.KoLConstants;
 import net.sourceforge.kolmafia.KoLConstants.MafiaState;
 import net.sourceforge.kolmafia.KoLmafia;
 import net.sourceforge.kolmafia.RequestThread;
-import net.sourceforge.kolmafia.ShopRow;
 import net.sourceforge.kolmafia.persistence.ItemFinder;
 import net.sourceforge.kolmafia.request.CoinMasterRequest;
+import net.sourceforge.kolmafia.shop.ShopRow;
 
 public class CoinmasterCommand extends AbstractCommand {
   public CoinmasterCommand() {
